bug fix & new boards & architecture
bug fix:
In the board constructor, when the player figure was sent to the bench, both a player movement and tile movement sound was played. Now only the player movement sound is played.
Also, when the player was on the field and then selected, the player movement sound was played. Now this is not the case any more.

new boards:
Apparently, when you go to the website of the game, there is a new manual that you can download as a .pdf. In this manual there are other boards than those of my print out manual that was shipped with the actual real world game. The boards.csv is equipped with the new boards.

architecture:
All files that are not imediatelly relevant to the game are stored inside the "more"-folder.